% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logging.R
\name{AssignLogFileToConfig}
\alias{AssignLogFileToConfig}
\title{Adds a log file to a \code{config} object.}
\usage{
AssignLogFileToConfig(file, config)
}
\arguments{
\item{file}{Character vector specifying path to log file.}

\item{config}{Configuration object created by LoadConfiguration.}
}
\description{
Adds a log file to a \code{config} object.
}
\examples{
\dontrun{
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie/global_config.yaml")
 config <- AssignLogFileToConfig(file = "~/nwisnfie/logs/myLog.log", config = config)
}
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

