% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{Autovacuum}
\alias{Autovacuum}
\alias{DisableAutovacuum}
\alias{EnableAutoVacuum}
\title{Manage Autovacuum Settings}
\usage{
EnableAutoVacuum(table, config)

DisableAutovacuum(table, config)
}
\arguments{
\item{table}{Character vector specifying the name of the table to modify.}

\item{config}{Configuration object created by LoadConfiguration.}
}
\description{
Functions used to enable and disable autovacuum on the data table
}
\details{
\code{EnableAutovacuum} enables autovacuum on the data table specified in \code{config}.
\code{DisableAutovacuum} disables autovacuum on the data table specified in \code{config}.
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

