% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dynamic_tables.R
\name{DynamicTables}
\alias{BuildDynamicTables}
\alias{DropDynamicTables}
\alias{DynamicTables}
\alias{RebuildDynamicTables}
\title{Manage dynamic tables.}
\usage{
BuildDynamicTables(config)

DropDynamicTables(config)

RebuildDynamicTables(config)
}
\arguments{
\item{config}{A configuration object created by LoadConfiguration}
}
\description{
The following tables are considered 'dynamic':
\enumerate{
\item data
}
}
\examples{
\dontrun{
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie/global_config.yaml")
 BuildDynamicTables(config)
 DropDynamicTables(config)
 # Equivalently...
 RebuildDynamicTables(config)
 }
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

