% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{Manage Database Connections}
\alias{Manage Database Connections}
\title{Manage Database Connections}
\arguments{
\item{conn}{A RPostgreSQL connection object.}

\item{config}{Configuration object created by LoadConfiguration}
}
\value{
For \code{StartDBConnection}, a RPostgreSQL connection object.
}
\description{
Functions used to start and stop database connections.
}
\examples{
\dontrun{
# Start and End Database Connection
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie/global_config.yaml")
 conn <- StartDBConnection(config)
 EndDBConnection(conn)
}
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

