\name{nwisnfie-package}
\alias{nwisnfie-package}
\alias{nwisnfie}
\docType{package}
\title{
National Water Information System - National Flood Interoperability Experiment
}
\description{
Facilitates data harvest from USGS National Water Information System and subsequent publication of data in NetCDF format. 
}
\details{
\tabular{ll}{
Package: \tab nwisnfie\cr
Type: \tab Package\cr
Version: \tab 0.1 \cr
Date: \tab 2015-01-05 \cr
License: \tab GPL (>= 2) \cr
}
}
\author{
Kevin M. Smith (https://github.com/Kevin-M-Smith)

Maintainer: Kevin.Smith@tufts.edu
}

\references{
\tabular{ll}{
NWIS: \tab http://waterdata.usgs.gov/nwis \cr
NetCDF:\tab http://www.unidata.ucar.edu/software/netcdf/ \cr
}
}

\keyword{ package }

