% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallel.R
\name{Parallelization}
\alias{Parallelization}
\alias{StartupCluster}
\alias{StopCluster}
\title{Cluster Management}
\usage{
StartupCluster(config)

StopCluster(cluster)
}
\arguments{
\item{config}{Configuration object created by \code{LoadConfiguration}}

\item{cluster}{A cluster reference produced by \code{StartupCluster}}
}
\value{
\code{StartupCluster} returns an object of class c("SOCKcluster", "cluster"), as built by \code{parallel::makePSOCKcluster}.
}
\description{
Functions for cleanly starting and stopping PSOCK clusters.
These clusters are used by \code{nwisnfie} to perform embarassingly parallel tasks in tandem.
PSOCK clusters are used for maximum compatibility across platforms.
}
\details{
\code{StartupCluster} starts up the cluster, loads the \code{nwisnfie} library, and starts individual database connections.
\code{StopCluster} closes out all database connections, and destroys the cluster.
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

