% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/static_tables.R
\name{PopulateStaticTables}
\alias{PopulateStaticTables}
\title{Populates static tables in database.}
\usage{
PopulateStaticTables(config)
}
\arguments{
\item{config}{A configuration object created by LoadConfiguration}
}
\description{
The following tables are considered 'static':
\enumerate{
\item param.codes
\item site.metadata
\item param.metadata
\item active.sites
\item site.assets
}
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

