% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{RunDBDiagnostics}
\alias{RunDBDiagnostics}
\title{Run Diagnotistics on Database}
\usage{
RunDBDiagnostics(config)
}
\arguments{
\item{config}{A configuration object created by LoadConfiguration}
}
\description{
The following tests are currently run:
\enumerate{
\item Load Database Connection
\item Check if required tables exist.
}
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

