% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{RunQuery}
\alias{RunQuery}
\title{Runs specified query.}
\usage{
RunQuery(conn, query, config)
}
\arguments{
\item{conn}{A RPostgreSQL connection object.}

\item{query}{A charactor vector of a PostgreSQL query.}

\item{config}{Configuration object created by LoadConfiguration.}
}
\value{
Result of query.
}
\description{
Runs specified query.
}
\examples{
\dontrun{
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie/global_config.yaml")
 conn <- StartDBConnection(config)
 result <- RunQuery(conn, query = "select * from data limit 1;", config = config)
 print(result)
 EndDBConnection(conn)
}
}
\seealso{
To build a \code{conn} object, see \code{\link{StartDBConnection}}.

To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

