% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/static_tables.R
\name{StaticTables}
\alias{DropStaticTables}
\alias{PopulateStaticTables}
\alias{RebuildStaticTables}
\alias{StaticTables}
\alias{WhichStaticTablesExist}
\title{Manage static tables.}
\usage{
PopulateStaticTables(config)

DropStaticTables(config)

WhichStaticTablesExist(config)

RebuildStaticTables(config)
}
\arguments{
\item{config}{A configuration object created by LoadConfiguration}
}
\description{
The following tables are considered 'static':
\enumerate{
\item param.codes
\item site.metadata
\item param.metadata
\item active.sites
\item site.assets
}
}
\examples{
\dontrun{
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie/global_config.yaml")
 PopulateStaticTables(config)
 DropStaticTables(config)
 # Equivalently...
 RebuildStaticTables(config)
 }
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

