% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{WhichTablesExist}
\alias{WhichTablesExist}
\title{Checks which tables exist.}
\usage{
WhichTablesExist(conn, config)
}
\arguments{
\item{conn}{Database connection created by StartDBConnection.}

\item{config}{A configuraiton object created by LoadConfiguration}
}
\description{
Checks for all tables listed in \code{\link{Configuration File}}.
}
\seealso{
To build a \code{conn} object, see \code{\link{StartDBConnection}}.
}

