% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{DBConnections}
\alias{DBConnections}
\alias{StartDBConnection}
\alias{StopDBConnection}
\alias{TestDBConnection}
\title{Manage Database Connections}
\usage{
StartDBConnection(config)

StopDBConnection(conn, config)

TestDBConnection(config)
}
\arguments{
\item{config}{Configuration object created by LoadConfiguration.}

\item{conn}{A RPostgreSQL connection object.}
}
\value{
For \code{StartDBConnection}, a RPostgreSQL connection object.
}
\description{
Functions used to start and stop database connections.
}
\details{
\code{StartDBConnection} starts a database connection specified by \code{config}.
\code{StopDBConnection} closes out the database connection contained in the \code{conn} object.
\code{TestDBConnection} attempts to start and stop a database connection based on \code{config}.
\dontrun{
# Start and End Database Connection
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie/global_config.yaml")
 conn <- StartDBConnection(config)
 StopDBConnection(conn)
# or equivalently:
 TestDBConnection(config)
}
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

