% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parallel.R
\name{Parallelization}
\alias{Parallelization}
\alias{StartCluster}
\alias{StopCluster}
\alias{TestClusterSettings}
\title{Cluster Management}
\usage{
StartCluster(config, ncores)

StopCluster(cluster, config)

TestClusterSettings(config)
}
\arguments{
\item{config}{Configuration object created by \code{LoadConfiguration}}

\item{cluster}{A cluster reference produced by \code{StartCluster}}
}
\value{
\code{StartCluster} returns an object of class c("SOCKcluster", "cluster"), as built by \code{parallel::makePSOCKcluster}.
}
\description{
Functions for cleanly starting and stopping PSOCK clusters.
These clusters are used by \code{nwisnfie} to perform embarassingly parallel tasks in tandem.
PSOCK clusters are used for maximum compatibility across platforms.
}
\details{
\code{StartCluster} starts up the cluster, loads the \code{nwisnfie} library, and starts individual database connections.
\code{StopCluster} closes out all database connections, and destroys the cluster.
\code{TestClusterSettings} runs \code{StartCluster} and \code{StopCluster} in succession.
}
\examples{
\dontrun{
 library(nwisnfie)
 config <- LoadConfiguration("~/nwisnfie/global_config.yaml")
 cluster <- StartCluster(config)
 StopCluster(cluster = cluster, config = config)
 # or equivalently:
 TestClusterSettings(config)
}
}
\seealso{
To build a \code{config} object, see \code{\link{LoadConfiguration}}.
}

